+++ src/features/lookingForParts/components/LookingForParts.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { PostCard } from "./PostCard";
+import { useLookingForParts } from "../hooks/useLookingForParts";
+
+const LookingForParts = () => {
+  const { play, handleAudioControl } = useLookingForParts();
+  return <PostCard play={play} handleAudioControl={handleAudioControl} />;
+};
+
+export default LookingForParts;
