+++ src/features/lookingForParts/hooks/useLookingForParts.ts
@@ -1,0 +1,14 @@
+import { useState } from "react";
+
+export const useLookingForParts = () => {
+  const [play, setPlay] = useState(false);
+
+  /**
+   * 再生ボタンクリック時にオーディオをコントロールする
+   */
+  const handleAudioControl = () => {
+    setPlay(!play);
+  };
+
+  return { play, handleAudioControl };
+};
