+++ src/features/lookingForParts/types/index.ts
@@ -1,0 +1,13 @@
+/**
+ * AudioControlのPropsの型定義
+ */
+export type AudioControlProps = {
+  /**
+   * オーディオの再生状態
+   */
+  play: boolean;
+  /**
+   * オーディオコントロールクリック時のイベント関数
+   */
+  handleAudioControl: () => void;
+};
